+++ run_device_tests.sh
@@ -49,3 +49,7 @@ cargo test test_reinit_duplex_stream_by_unplugging_a_default_input_device -- --i
 
 cargo test test_destroy_duplex_stream_after_unplugging_a_default_output_device -- --ignored --nocapture
 cargo test test_reinit_duplex_stream_by_unplugging_a_default_output_device -- --ignored --nocapture
+
+cargo test test_device_collection_change_on_default_input_device_change -- --ignored --nocapture
+cargo test test_device_collection_change_on_default_output_device_change -- --ignored --nocapture
+cargo test test_device_collection_change_on_default_device_change_duplex -- --ignored --nocapture
