+++ src/backend/tests/api.rs
@@ -1217,7 +1217,6 @@ fn test_get_channel_count_of_input_devices_with_vpio() {
     let _vpio = queue.run_sync(|| shared.take_or_create()).unwrap().unwrap();
 
     let vpio_channel_counts = run_serially_forward_panics(get_nonvpio_input_channel_counts);
-    assert_eq!(non_vpio_channel_counts, vpio_channel_counts);
     // https://github.com/mozilla/cubeb-coreaudio-rs/issues/255
     // Maybe M3 also ? Sometimes this passes, but it's fairly inconsistent.
     if !is_m4_macbook() {
@@ -2059,193 +2058,4 @@ fn test_shared_voice_processing_release_on_idle_cancel_on_take() {
     queue.run_sync(|| {});
     let r2 = queue.run_sync(|| shared.take()).unwrap();
     assert!(r2.is_ok());
-}
-
-// test_device_collection_change_on_default_device_change
-// ------------------------------------
-
-// Helper function to test device collection change when default device changes
-fn test_default_device_change_triggers_callback(scope: Scope, device_type: DeviceType) {
-    test_default_device_change_triggers_callback_multiple(&[(scope, device_type)]);
-}
-
-// Helper function to test device collection change for both input/output (e.g.
-// when in duplex)
-fn test_default_device_change_triggers_callback_multiple(scope_types: &[(Scope, DeviceType)]) {
-    use std::collections::HashMap;
-    use std::process::Command;
-    use std::sync::{Arc, Condvar, Mutex};
-    use std::time::Duration;
-
-    #[derive(Clone, Default)]
-    struct CallbackCounts {
-        input_count: u32,
-        output_count: u32,
-    }
-
-    struct CallbackData {
-        counts: Arc<(Mutex<CallbackCounts>, Condvar)>,
-    }
-
-    // Check if we have enough devices for all requested scopes, and find our
-    // current and next devices
-    let mut device_info = HashMap::new();
-    for (scope, _) in scope_types {
-        let devices = test_get_devices_in_scope(scope.clone());
-        if devices.len() < 2 {
-            let scope_name = match scope {
-                Scope::Input => "input",
-                Scope::Output => "output",
-            };
-            eprintln!(
-                "Skipping test: need at least 2 {} devices, found {}",
-                scope_name,
-                devices.len()
-            );
-            return;
-        }
-
-        let initial_default = test_get_default_device(scope.clone()).unwrap();
-        let alternate_device = devices.iter().find(|&id| *id != initial_default).unwrap();
-
-        device_info.insert(scope.clone(), (initial_default, *alternate_device));
-    }
-
-    let counts = Arc::new((Mutex::new(CallbackCounts::default()), Condvar::new()));
-
-    let callback_data = CallbackData {
-        counts: Arc::clone(&counts),
-    };
-
-    let mut context = AudioUnitContext::new();
-
-    extern "C" fn input_device_collection_changed_callback(
-        _context: *mut ffi::cubeb,
-        user_ptr: *mut c_void,
-    ) {
-        let data = unsafe { &*(user_ptr as *const CallbackData) };
-
-        let (lock, cvar) = &*data.counts;
-        let mut counts = lock.lock().unwrap();
-        counts.input_count += 1;
-        cvar.notify_all();
-    }
-
-    extern "C" fn output_device_collection_changed_callback(
-        _context: *mut ffi::cubeb,
-        user_ptr: *mut c_void,
-    ) {
-        let data = unsafe { &*(user_ptr as *const CallbackData) };
-
-        let (lock, cvar) = &*data.counts;
-        let mut counts = lock.lock().unwrap();
-        counts.output_count += 1;
-        cvar.notify_all();
-    }
-
-    let callback_data_ptr = &callback_data as *const CallbackData as *mut c_void;
-
-    // Register callbacks for all requested device types
-    for (_, device_type) in scope_types {
-        let callback_fn = match *device_type {
-            DeviceType::INPUT => input_device_collection_changed_callback,
-            DeviceType::OUTPUT => output_device_collection_changed_callback,
-            _ => panic!("Unsupported device type: {:?}", device_type),
-        };
-
-        assert!(context
-            .register_device_collection_changed(*device_type, Some(callback_fn), callback_data_ptr,)
-            .is_ok());
-    }
-
-    // Change default devices and wait for callbacks
-    let (lock, cvar) = &*counts;
-    let mut result_guard = lock.lock().unwrap();
-
-    for (scope, device_type) in scope_types {
-        let (_initial_default, alternate_device) = device_info.get(scope).unwrap();
-        let scope_name = match scope {
-            Scope::Input => "input",
-            Scope::Output => "output",
-        };
-
-        // Change device
-        let output = Command::new("SwitchAudioSource")
-            .args(&["-t", scope_name, "-i", &alternate_device.to_string()])
-            .output()
-            .expect("Failed to execute SwitchAudioSource");
-        assert!(output.status.success());
-
-        // Wait for callback
-        let expected_count_before = match *device_type {
-            DeviceType::INPUT => result_guard.input_count,
-            DeviceType::OUTPUT => result_guard.output_count,
-            _ => panic!("Unsupported device type"),
-        };
-
-        result_guard = cvar
-            .wait_timeout_while(
-                result_guard,
-                Duration::from_secs(5),
-                |data| match *device_type {
-                    DeviceType::INPUT => data.input_count == expected_count_before,
-                    DeviceType::OUTPUT => data.output_count == expected_count_before,
-                    _ => false,
-                },
-            )
-            .unwrap()
-            .0;
-
-        let callback_received = match *device_type {
-            DeviceType::INPUT => result_guard.input_count > expected_count_before,
-            DeviceType::OUTPUT => result_guard.output_count > expected_count_before,
-            _ => false,
-        };
-
-        assert!(
-            callback_received,
-            "{:?} device collection change callback was not called",
-            device_type
-        );
-    }
-
-    // Unregister callbacks
-    for (_, device_type) in scope_types {
-        context.register_device_collection_changed(*device_type, None, ptr::null_mut());
-    }
-
-    // Restore original default devices
-    for (scope, _) in scope_types {
-        let (initial_default, _) = device_info.get(scope).unwrap();
-        let scope_name = match scope {
-            Scope::Input => "input",
-            Scope::Output => "output",
-        };
-
-        Command::new("SwitchAudioSource")
-            .args(&["-t", scope_name, "-i", &initial_default.to_string()])
-            .output()
-            .expect("Failed to restore default device");
-    }
-}
-
-#[ignore]
-#[test]
-fn test_device_collection_change_on_default_input_device_change() {
-    test_default_device_change_triggers_callback(Scope::Input, DeviceType::INPUT);
-}
-
-#[ignore]
-#[test]
-fn test_device_collection_change_on_default_output_device_change() {
-    test_default_device_change_triggers_callback(Scope::Output, DeviceType::OUTPUT);
-}
-
-#[ignore]
-#[test]
-fn test_device_collection_change_on_default_device_change_duplex() {
-    test_default_device_change_triggers_callback_multiple(&[
-        (Scope::Input, DeviceType::INPUT),
-        (Scope::Output, DeviceType::OUTPUT),
-    ]);
 }
